gcc: Additional warning cleanup
This commit addresses some issues encountered when adding the
`-Wconversion` and `-Wno-sign-conversion` flags for stricter type
and sign conversion checks.

It also removes the `ZSL_MTX_UNARY_OP_LOGICAL_NEGATION` unary operand
in matrices.s since it is of marginal use in the real world, and causes
a cast issue from float to bool.

+++ src/vectors.c
@@ -110,7 +110,7 @@ int zsl_vec_neg(struct zsl_vec *v)
 
 int zsl_vec_sum(struct zsl_vec **v, size_t n, struct zsl_vec *w)
 {
-	int sz_last;
+	size_t sz_last;
 
 	if (!n) {
 		return -EINVAL;
